fix:delete CountDownHub
