+++ Components/_Imports.razor
@@ -13,3 +13,6 @@
 @using G5_MovieTicketBookingSystem.Components.Layout
 @using G5_MovieTicketBookingSystem.Components.Shared
 @using G5_MovieTicketBookingSystem.Components.Pages
+@using G5_MovieTicketBookingSystem.DTOs.VNPAY
+@using G5_MovieTicketBookingSystem.Repositories
+@using G5_MovieTicketBookingSystem.Services.Impl
