--- Models/VNPAY/PaymentResponseModel.cs
@@ -1,15 +1,0 @@
-﻿namespace G5_MovieTicketBookingSystem.Models.VNPAY
-{
-    public class PaymentResponseModel
-    {
-          public string OrderDescription { get; set; }
-    public string TransactionId { get; set; }
-    public string OrderId { get; set; }
-    public string PaymentMethod { get; set; }
-    public string PaymentId { get; set; }
-    public bool Success { get; set; }
-    public string Token { get; set; }
-    public string VnPayResponseCode { get; set; }
-        public string PaymentMessage { get; internal set; }
-    }
-}
