+++ Services/Impl/PaymentService.cs
@@ -1,5 +1,5 @@
 ﻿using G5_MovieTicketBookingSystem.Data;
-using G5_MovieTicketBookingSystem.Models;
+//using G5_MovieTicketBookingSystem.Models;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Linq;
