+++ Services/Impl/SeatLockService.cs
@@ -1,17 +1,14 @@
 ﻿using G5_MovieTicketBookingSystem.Repositories;
 using G5_MovieTicketBookingSystem;
 using Microsoft.AspNetCore.SignalR;
-using G5_MovieTicketBookingSystem.Hubs;
 
 public class SeatLockService : ISeatLockService
 {
     private readonly ISeatLockRepository _seatLockRepository;
-    private readonly IHubContext<CountdownHub> _hubContext;
 
-    public SeatLockService(ISeatLockRepository seatLockRepository, IHubContext<CountdownHub> hubContext)
+    public SeatLockService(ISeatLockRepository seatLockRepository)
     {
         _seatLockRepository = seatLockRepository;
-        _hubContext = hubContext;
     }
 
     public async Task<SeatLock?> GetLatestSeatLockByUserIdAsync(int? userId)
@@ -20,8 +17,5 @@ public class SeatLockService : ISeatLockService
         return seatLock;
     }
 
-    public async Task SendTimeToClients(string timeRemaining)
-    {
-        await _hubContext.Clients.All.SendAsync("ReceiveTime", timeRemaining);
-    }
+ 
 }
