+++ Services/Impl/VnPayService.cs
@@ -1,5 +1,5 @@
-﻿using G5_MovieTicketBookingSystem.libraries;
-using G5_MovieTicketBookingSystem.Models.VNPAY;
+﻿using G5_MovieTicketBookingSystem.DTOs.VNPAY;
+using G5_MovieTicketBookingSystem.libraries;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using System;
